Update xkablevif() function
xkablevif() was putting out different results from vif() when the regressors were not in alphabetical order. The table() function reordered the names of the coefficients to alphabetical order but the values that were being assigned to vifs object were still in the order they were input into lm(). I added one step to sort the vif results by the coefficient names before setting the values.